+++ src/commands/deploy.rs
@@ -6,8 +6,8 @@ use crate::{
     config::TargetConfig,
     logger,
     project::{load_project, Project},
-    safe_resource_manager::{RobloxInputs, RobloxOutputs, RobloxResource, RobloxResourceManager},
-    safe_resources::{EvaluateResults, ResourceGraph},
+    roblox_resource_manager::{RobloxInputs, RobloxOutputs, RobloxResource, RobloxResourceManager},
+    resource_graph::{EvaluateResults, ResourceGraph},
     state::save_state,
     util::run_command,
 };
