+++ src/state/v3.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 use serde::{Deserialize, Serialize};
 
-use crate::safe_resource_manager::RobloxResource;
+use crate::roblox_resource_manager::RobloxResource;
 
 #[derive(Serialize, Deserialize, Clone)]
 pub struct ResourceStateV3 {
